requirements.txt
streamlit
langchain
pdf2image
requests
python-dotenv
langchain-google-genai